refactor(struggle): neutral session-lifecycle module, buildResultGuard into sensing
