+++ extension/src/extension/services/sensing/buildResultGuard.ts
